@@ -2,13 +2,14 @@ source 'https://rubygems.org'
 
 gem 'rake', '~> 10.4.2'
 gem 'rack', '2.0.1'
-gem 'rails', '5.0.1'
+gem 'rails', '5.1.3'
 #gem 'jbuilder'
 
 gem 'pg'
 gem 'sqlite3'
 
 gem 'config'
+gem 'jbuilder'
 
 #gem 'sass-rails', '~> 4.0.3'
 gem 'sassc-rails', '~> 1.3.0'
@@ -106,7 +107,8 @@ end
 gem 'gi_test_engine', git: 'git@github.com:kikonen/gi_test_engine.git', branch: 'master'
 #gem 'gi_test_engine', path: '~/work/projects/ruby/gi_test_engine'
 
-gem 'gi_album_engine', git: 'git@github.com:kikonen/gi_album_engine.git', branch: 'real_album'
+gem 'gi_album_engine', git: 'git@github.com:kikonen/gi_album_engine.git', branch: 'master'
+#gem 'gi_album_engine', git: 'git@github.com:kikonen/gi_album_engine.git', branch: 'real_album'
 #gem 'gi_album_engine', path: '~/work/projects/ruby/gi_album_engine'
 
 gem 'gi_raycaster_engine', git: 'git@github.com:kikonen/gi_raycaster_engine.git', branch: 'master'
