@@ -3,7 +3,7 @@ class ApplicationController < ActionController::Base
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
 
-  before_filter :require_theme
+  before_action :require_theme
 
   def require_theme
     @theme_id = 'cerulean'
