@@ -3,19 +3,23 @@ ENV['BUNDLE_GEMFILE'] ||= File.expand_path('../../Gemfile', __FILE__)
 
 require 'bundler/setup' if File.exist?(ENV['BUNDLE_GEMFILE'])
 
-require 'rails/commands/server'
-
 # No stdout for logger
+require 'rack'
+require 'rails/command'
+require 'rails/commands/server/server_command'
+
 module Rails
   class Server < ::Rack::Server
-    class Options
-      alias_method :orig_parse, :parse!
+    alias_method :orig_start, :start
 
-      def parse!(args)
-        options = orig_parse(args)
-        options[:log_stdout] = false
-        options
-      end
+    def start
+      orig_start
+      options[:log_stdout] = false
+    end
+
+    private
+    def log_to_stdout
+      # No stdout for logger
     end
   end
 end
